Moved listener to a new subproject: paper
