+++ paper/src/main/java/net/valneas/account/listener/RankAddedListener.java
@@ -1,30 +1,29 @@
 package net.valneas.account.listener;
 
 import io.github.leonardosnt.bungeechannelapi.BungeeChannelApi;
-import net.valneas.account.AccountManager;
 import net.valneas.account.AccountSystem;
-import net.valneas.account.api.events.rank.RankRemovedEvent;
+import net.valneas.account.events.rank.RankAddedEvent;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 
-public class RankRemovedListener implements Listener {
+public class RankAddedListener implements Listener {
 
     private final AccountSystem main;
 
-    public RankRemovedListener(AccountSystem main) {
+    public RankAddedListener(AccountSystem main) {
         this.main = main;
     }
 
     @EventHandler
-    public void onRankRemoved(RankRemovedEvent e){
-        final AccountManager account = e.getAccount();
+    public void onRankAdded(RankAddedEvent e){
+        final var account = e.getAccount();
 
         String message = ChatColor.YELLOW + "" + ChatColor.ITALIC + "Michel §r§l➔ " + ChatColor.YELLOW +
-                "Coucou, apparement tu as perdu le rang §r" + ChatColor.BOLD + e.getRankRemoved().getName() + ChatColor.YELLOW +
-                ". Je ne ressens pas d'émotion mais si c'était important, je suis désolé pour toi. #SAD";
+                "Youpiii on vient de t'ajouter le rang §r" + ChatColor.BOLD + e.getRankAdded().getName() + ChatColor.YELLOW +
+                "! Profitez-en bien";
 
         if(main.isBungeeMode()){
             final BungeeChannelApi api = BungeeChannelApi.of(main);
